3375: Move tabs to separate component
